- Further progress on parallel kmedoids implementation.  Now works up to broadcast phase; still need to implement parallel medoid assignment and cluster criterion. - Added debug output to kmedoids pam implementation to debug convergence problem with input where two sets of clusters have equal cluster quality.

+++ random.h
@@ -17,7 +17,7 @@ template <class OutputIterator, class Random>
 void random_subset(size_t numElements, size_t sample_size, OutputIterator out, Random& random) {
   size_t first = 0;
   size_t remaining = numElements;
-  long long m = sample_size;
+  size_t m = sample_size;
   
   while (m > 0) {
     if (random(remaining) < m) {
